requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0
jinja2==3.1.6
python-multipart==0.0.26
pydantic-settings==2.2.1
